BUG: Need to change the return type of GetMin/GetMax to support floating point image

+++ Source/MediaStorageAndFileFormat/gdcmPixelFormat.h
@@ -130,8 +130,8 @@ public:
 
   void Print(std::ostream &os) const;
 
-  int64_t GetMin() const;
-  int64_t GetMax() const;
+  double GetMin() const;
+  double GetMax() const;
 
 private:
   // D 0028|0002 [US] [Samples per Pixel] [1]
